refactor(core): move OffsetPagination to _pagination.py (mypyc compat) (#419)
Discovered during wheel validation (C2.T5): mypyc refuses to compile any
class that declares a metaclass, and msgspec.Struct uses one. Colocating
OffsetPagination as msgspec.Struct inside the mypyc-compiled
sqlspec/core/filters.py breaks wheel builds with:

  TypeError: mypyc classes can't have a metaclass

Fix: split OffsetPagination into a sibling module sqlspec/core/_pagination.py
and add that file to the mypyc exclude list. Re-export from
sqlspec.core.filters so the public import path is unchanged. The hot-path
filter classes (StatementFilter, BeforeAfterFilter, etc.) remain in
filters.py and continue to be compiled.

Also switched the OpenAPISchemaPlugin from string 'array'/'object'/'integer'
Schema types to litestar.openapi.spec.OpenAPIType enum members to satisfy
Litestar's type signature.

Verified: compiled wheel loads, OffsetPagination.__annotations__ is
non-empty, #419 reproducer script exits with 'OK: Item registered, response
schema non-empty'.

Beads: sqlspec-p9d.2.2, sqlspec-p9d.2.5
Refs: #419
